Add tab completion for most commands
